refactor: move file for ownership
