+++ public/app/features/auth-config/types.ts
@@ -1,4 +1,4 @@
-import { AuthProviderStatus } from 'app/types';
+import { Settings } from 'app/types';
 
 export interface AuthProviderInfo {
   id: string;
@@ -9,3 +9,23 @@ export interface AuthProviderInfo {
 }
 
 export type GetStatusHook = () => Promise<AuthProviderStatus>;
+
+export interface AuthConfigState {
+  settings: Settings;
+  providerStatuses: Record<string, AuthProviderStatus>;
+  isLoading?: boolean;
+  updateError?: SettingsError;
+  warning?: SettingsError;
+}
+
+export interface AuthProviderStatus {
+  enabled: boolean;
+  configured: boolean;
+  configFoundInIniFile?: boolean;
+  hide?: boolean;
+}
+
+export interface SettingsError {
+  message: string;
+  errors: string[];
+}
