+++ public/app/types/index.ts
@@ -18,7 +18,6 @@ export * from './appEvent';
 export * from './query';
 export * from './accessControl';
 export * from './supportBundles';
-export * from './configAuth';
 export * from './settings';
 
 import * as CoreEvents from './events';
